==> 01_dataCleaning_09_feat_homeGame.R <==

# CREATING HOMEGAME FEATURE FOR D

# Setup --------------------------------------------------------------

library(tidyverse)

rm(list = ls())

# Reading the previously saved version of our data
load("../Roeser, Jonas - 2_Data/D1.RData")

# Because of OneDrive we need to load from two different paths
load("../2_Data/D1.RData")


# Creating homeGames feature  ---------------------------------------------------

D1 = D1[c(2,7,24,42)] %>%
  mutate(home_game_p0 = 0) %>%
  mutate(home_game_p1 = 0)

D1$tourney_location = as.character(D1$tourney_location)

unique(D1$tourney_location)

D1 <- D1 %>%
  mutate(tourney_location = sapply(D1$tourney_location, switch,
                                   "Melbourne" = "AUS",
                                   "Guaruja" = "BRA",
                                   "Milan" = "ITA",
                                   "San Francisco" = "USA",
                                   "Brussels" = "BEL",
                                   "Philadelphia" = "USA",
                                   "Memphis" = "USA",
                                   "Stuttgart" = "GER",
                                   "Chicago" = "USA",
                                   "Rotterdam" = "NED",
                                   "Copenhagen" = "DEN",
                                   "Indian Wells" = "USA",
                                   "Miami" = "USA",
                                   "Estoril" = "POR",
                                   "Hong Kong" = "HKG",
                                   "Orlando" = "USA",
                                   "Barcelona" = "ESP",
                                   "Tokyo" = "JPN",
                                   "Singapore" = "SIN",
                                   "Madrid" = "ESP",
                                   "Munich" = "GER",
                                   "Tampa" = "USA",
                                   "Charlotte" = "USA",
                                   "Hamburg" = "GER",
                                   "Rome" = "ITA",
                                   "Umag" = "CRO",
                                   "Bologna" = "ITA",
                                   "Rosmalen" = "NED",
                                   "Gstaad" = "SUI",
                                   "Montreal" = "CAN",
                                   "Cincinnati" = "USA",
                                   "Long Island" = "USA",
                                   "Schenectady" = "USA",
                                   "New York" = "USA",
                                   "Bordeaux" = "FRA",
                                   "Brasilia" = "BRA",
                                   "Geneva" = "SUI",
                                   "Palermo" = "ITA",
                                   "Athens" = "GRE",
                                   "Sydney" = "AUS",
                                   "Toulouse" = "FRA",
                                   "Berlin" = "GER",
                                   "Tel Aviv" = "ISR",
                                   "Lyon" = "FRA",
                                   "Vienna" = "AUT",
                                   "Stockholm" = "SWE",
                                   "Buzios" = "BRA",
                                   "Birmingham" = "GBR",
                                   "Moscow" = "RUS",
                                   "Sao Paulo" = "BRA",
                                   "Adelaide" = "AUS",
                                   "Wellington" = "NZL",
                                   "Auckland" = "NZL",
                                   "Maceio" = "BRA",
                                   "Scottsdale" = "USA",
                                   "Casablanca" = "MAR",
                                   "Johannesburg" = "RSA",
                                   "Atlanta" = "USA",
                                   "Toronto" = "CAN",
                                   "Cologne" = "GER",
                                   "Bolzano" = "ITA",
                                   "Taipei" = "TPE",
                                   "Antwerp" = "BEL",
                                   "Doha" = "QAT",
                                   "Kuala Lumpur" = "MAS",
                                   "Jakarta" = "INA",
                                   "Dubai" = "UAE",
                                   "Marseille" = "FRA",
                                   "Mexico City" = "MEX",
                                   "Zaragoza" = "ESP",
                                   "Durban" = "RSA",
                                   "Osaka" = "JPN",
                                   "Delray Beach" = "USA",
                                   "Halle" = "GER",
                                   "Bucharest" = "ROU",
                                   "Beijing" = "CHN",
                                   "Santiago" = "MEX",
                                   "Buenos Aires" = "ARG",
                                   "Oahu" = "USA",
                                   "San Jose" = "CAC",
                                   "Sun City" = "RSA",
                                   "St. Poelten" = "AUT",
                                   "Bogota" = "COL",
                                   "Ostrava" = "CZE",
                                   "Montevideo" = "URU",
                                   "St. Petersburg" = "RUS",
                                   "Bermuda" = "BER",
                                   "Pinehurst" = "USA",
                                   "Oporto" = "POR",
                                   "Essen" = "GER",
                                   "Shanghai" = "CHN",
                                   "Zagreb" = "CRO",
                                   "New Delhi" = "IND",
                                   "Bournemouth" = "GBR",
                                   "Marbella" = "ESP",
                                   "Chennai" = "CHN",
                                   "Boston" = "USA",
                                   "Tashkent" = "UZB",
                                   "Split" = "CRO",
                                   "s-Hertogenbosch" = "NED",
                                   "Kitzbuehel" = "AUT",
                                   "Mallorca" = "ESP",
                                   "Merano" = "ITA",
                                   "Shanghai, China" = "CHN",
                                   "Brighton" = "GBR",
                                   "Vina Del Mar" = "CHI",
                                   "Acapulco" = "MEX",
                                   "Houston" = "USA",
                                   "Sopot" = "POL",
                                   "Amersfoort" = "NED",
                                   "Bangkok" = "THA",
                                   "Metz" = "FRA",
                                   "Ho Chi Minh City" = "VIE",
                                   "Las Vegas" = "USA",
                                   "Poertschach" = "AUT",
                                   "Mumbai" = "IND",
                                   "Warsaw" = "POL",
                                   "Kitzbhel" = "AUT",
                                   "Acapulco, Mexico" = "MEX",
                                   "Belgrade" = "SRB",
                                   "Eastbourne" = "GBR",
                                   "Bucharest, Romania" = "ROU",
                                   "Marseille, France" = "FRA",
                                   "Dubai, United Arab Emirates" = "UAE",
                                   "Houston, United States" = "USA",
                                   "Munich, Germany" = "GER",
                                   "Madrid, Spain" = "ESP",
                                   "Paris, France" = "FRA",
                                   "Halle, Germany" = "GER",
                                   "London, Great Britain" = "GBR",
                                   "Eastbourne, Great Britain" = "GBR",
                                   "s-Hertogenbosch, Netherlands" = "NED",
                                   "Newport, United States" = "USA",
                                   "Bastad, Sweden" = "SWE",
                                   "Stuttgart, Germany" = "GER",
                                   "Atlanta, United States" = "USA",
                                   "Hamburg, Germany" = "GER",
                                   "Gstaad, Switzerland" = "SUI",
                                   "Los Angeles, United States" = "USA",
                                   "Umag, Croatia" = "CRO",
                                   "Washington, United States" = "USA",
                                   "Toronto, Canada" = "CAN",
                                   "Cincinnati, United States" = "USA",
                                   "Beijing, China" = "CHN",
                                   "Moscow, Russia" = "RUS",
                                   "Montpellier, France" = "FRA",
                                   "St. Petersburg, Russia" = "RUS",
                                   "Vienna, Austria" = "AUT",
                                   "Valencia, Switzerland" = "ESP",
                                   "Basel, Switzerland" = "SUI",
                                   "Brisbane, Australia" = "AUS",
                                   "Chennai, India" = "IND",
                                   "Doha, Qatar" = "QAT",
                                   "Auckland, New Zealand" = "NZL",
                                   "Sydney, Australia" = "AUS",
                                   "Melbourne, Australia" = "AUS",
                                   "Johannesburg, South Africa" = "RSA",
                                   "Santiago, Chile" = "CHI",
                                   "Zagreb, Croatia" = "CRO",
                                   "Rotterdam, Netherlands" = "NED",
                                   "San Jose, United States" = "USA",
                                   "Sao Paulo, Brazil" = "BRA",
                                   "Buenos Aires, Argentina" = "ARG",
                                   "Memphis, United States" = "USA",
                                   "Delray Beach, United States" = "USA",
                                   "Indian Wells, United States" = "USA",
                                   "Miami, United States" = "USA",
                                   "Casablanca, Morocco" = "MOR",
                                   "Monte Carlo, Monaco" = "MON",
                                   "Barcelona, Spain" = "ESP",
                                   "Belgrade, Serbia" = "SRB",
                                   "Estoril, Portugal" = "POR",
                                   "Rome, Italy" = "ITA",
                                   "Kitzbhel, Austria" = "AUT",
                                   "Montreal, Canada" = "CAN",
                                   "Winston-Salem, United States" = "USA",
                                   "New York, United States" = "USA",
                                   "Metz, France" = "FRA",
                                   "Bangkok, Thailand" = "THA",
                                   "Kuala Lumpur, Malaysia" = "MAS",
                                   "Tokyo, Japan" = "JPN",
                                   "Stockholm, Sweden" = "SWE",
                                   "Valencia, Spain" = "ESP",
                                   "Vina del Mar, Chile" = "CHI",
                                   "Monaco, Monaco" = "MON",
                                   "Bâ€ stad, Sweden" = "SWE",
                                   "Kitzbuhel, Austria" = "AUT",
                                   "Monte-Carlo, Monaco" = "MON",
                                   "Oeiras, Portugal" = "POR",
                                   "Dusseldorf, Germany" = "GER",
                                   "Bogota, Colombia" = "COL",
                                   "Rio de Janeiro, Brazil" = "BRA",
                                   "New York City, United States" = "USA",
                                   "Shenzhen, China" = "CHN",
                                   "Quito, Ecuador" = "ECU",
                                   "Istanbul, Turkey" = "TUR",
                                   "Geneva, Switzerland" = "SUI",
                                   "Nottingham, Great Britain" = "GBR",
                                   "Sofia, Bulgaria" = "BUL",
                                   "Marrakech, Morocco" = "MOR",
                                   "Los Cabos, Mexico" = "MEX",
                                   "Chengdu, China" = "CHN",
                                   "Antwerp, Belgium" = "BEL",
                                   "Budapest, Hungary" = "HUN",
                                   "Lyon, France" = "FRA",
                                   "Antalya, Turkey" = "TUR"
                                   ))

unique(D1$tourney_location)

D1$home_game_p0[which(D1$tourney_location == D1$flag_code.x)] = 1
D1$home_game_p1[which(D1$tourney_location == D1$flag_code.y)] = 1


# Saving -------------------------------------------------------------

# Saving "D1" as "homeGame.RData"
homeGame = D1
# save(homeGame, file = "../Roeser, Jonas - 2_Data/homeGame.RData")
